Remove the socials from the signin page
